metasploit-framework.gemspec
[#48047663]

Create metasploit-framework.gemspec as a first attempt at turn
metasploit-framework into a gem that pro can require in its Gemfile
instead of manually setting the $LOAD_PATH to include pro/msf3.

Executables were moved out of the root into a bin directory to follow
gem conventions.  gemspec was added to Gemfile and ungrouped (default)
dependencies were added as runtime_dependencies to the
Gem::Specification.  Bundler::GemHelper.install_tasks was added to
Rakefile to add the gem rake tasks, such as build, install and release.

Other code change was moving the versioning information out of
Msf::Framework because it requires too much to be loaded (and can't be
required directly without error) for the .gemspec file to load the
versioning information out of Msf::Framework, so it is now in
Metasploit::Framework::Version.